chore: make site full width
