+++ site/src/app/(v2)/(blog)/changelog/layout.tsx
@@ -1,10 +1,10 @@
-import { FancyHeader } from "@/components/v2/FancyHeader";
+import { Header } from "@/components/v2/Header";
 import type { CSSProperties } from "react";
 
 export default function Layout({ children }) {
 	return (
 		<>
-			<FancyHeader active="blog" />
+			<Header active="blog" variant="floating" />
 			<div
 				className="mx-auto mt-20 w-full max-w-6xl px-8 md:mt-32"
 				style={{ "--header-height": "5rem" } as CSSProperties}
