+++ site/src/app/(v2)/(content)/layout.tsx
@@ -1,12 +1,12 @@
 import { Prose } from "@/components/Prose";
-import { FancyHeader } from "@/components/v2/FancyHeader";
+import { Header } from "@/components/v2/Header";
 import { cn } from "@rivet-gg/components";
 import type { CSSProperties } from "react";
 
 export default function Layout({ children }: { children: React.ReactNode }) {
 	return (
 		<>
-			<FancyHeader />
+			<Header variant="floating" />
 			<div
 				className="md:mt-32 mt-20 pb-8"
 				style={{ "--header-height": "5rem" } as CSSProperties}
