+++ site/src/app/(v2)/not-found.tsx
@@ -1,11 +1,11 @@
 import { Button } from "@/components/Button";
-import { FancyHeader } from "@/components/v2/FancyHeader";
+import { Header } from "@/components/v2/Header";
 import { faBlockQuestion, Icon } from "@rivet-gg/icons";
 
 export default function NotFound() {
 	return (
 		<>
-			<FancyHeader />
+			<Header variant="floating" />
 			<div className="relative flex min-h-[80vh] w-full items-center justify-center text-center">
 				<div className="transition-opacity">
 					<h1 className="mb-3 flex items-center justify-center text-3xl text-white">
